Linux 2.0.36pre-patch 20
o	PCI ID for a newer lance chip
o	TCP should now handle syn+ack+data frames right again (KA9Q sends them)
o	ISDN4linux update - this should fix some stack overruns with Hisax

+++ drivers/isdn/avmb1/b1lli.c
@@ -1,11 +1,14 @@
 /*
- * $Id: b1lli.c,v 1.1.2.10 1998/03/20 20:34:41 calle Exp $
+ * $Id: b1lli.c,v 1.1.2.11 1998/10/25 14:36:18 fritz Exp $
  * 
  * ISDN lowlevel-module for AVM B1-card.
  * 
  * (c) Copyright 1997 by Carsten Paeth (calle@calle.in-berlin.de)
  * 
  * $Log: b1lli.c,v $
+ * Revision 1.1.2.11  1998/10/25 14:36:18  fritz
+ * Backported from MIPS (Cobalt).
+ *
  * Revision 1.1.2.10  1998/03/20 20:34:41  calle
  * port valid check now only for T1, because of the PCI and PCMCIA cards.
  *
@@ -213,7 +216,7 @@ extern int showcapimsgs;
 #define B1_STAT1(cardtype)  (0x80E00000l)
 
 
-static inline unsigned char b1outp(unsigned short base,
+static inline unsigned char b1outp(unsigned int base,
 				   unsigned short offset,
 				   unsigned char value)
 {
@@ -221,34 +224,34 @@ static inline unsigned char b1outp(unsigned short base,
 	return inb(base + B1_ANALYSE);
 }
 
-static inline void t1outp(unsigned short base,
+static inline void t1outp(unsigned int base,
 			  unsigned short offset,
 			  unsigned char value)
 {
 	outb(value, base + offset);
 }
 
-static inline unsigned char t1inp(unsigned short base,
+static inline unsigned char t1inp(unsigned int base,
 			          unsigned short offset)
 {
 	return inb(base + offset);
 }
 
-static inline int B1_isfastlink(unsigned short base)
+static inline int B1_isfastlink(unsigned int base)
 {
 	return (inb(base + T1_IDENT) & ~0x82) == 1;
 }
-static inline unsigned char B1_fifostatus(unsigned short base)
+static inline unsigned char B1_fifostatus(unsigned int base)
 {
 	return inb(base + T1_FIFOSTAT);
 }
 
-static inline int B1_rx_full(unsigned short base)
+static inline int B1_rx_full(unsigned int base)
 {
 	return inb(base + B1_INSTAT) & 0x1;
 }
 
-static inline unsigned char B1_get_byte(unsigned short base)
+static inline unsigned char B1_get_byte(unsigned int base)
 {
 	unsigned long i = jiffies + 1 * HZ;	/* maximum wait time 1 sec */
 	while (!B1_rx_full(base) && i > jiffies);
@@ -258,7 +261,7 @@ static inline unsigned char B1_get_byte(unsigned short base)
 	return 0;
 }
 
-static inline unsigned int B1_get_word(unsigned short base)
+static inline unsigned int B1_get_word(unsigned int base)
 {
 	unsigned int val = 0;
 	val |= B1_get_byte(base);
@@ -268,18 +271,18 @@ static inline unsigned int B1_get_word(unsigned short base)
 	return val;
 }
 
-static inline int B1_tx_empty(unsigned short base)
+static inline int B1_tx_empty(unsigned int base)
 {
 	return inb(base + B1_OUTSTAT) & 0x1;
 }
 
-static inline void B1_put_byte(unsigned short base, unsigned char val)
+static inline void B1_put_byte(unsigned int base, unsigned char val)
 {
 	while (!B1_tx_empty(base));
 	b1outp(base, B1_WRITE, val);
 }
 
-static inline void B1_put_word(unsigned short base, unsigned int val)
+static inline void B1_put_word(unsigned int base, unsigned int val)
 {
 	B1_put_byte(base, val & 0xff);
 	B1_put_byte(base, (val >> 8) & 0xff);
@@ -287,7 +290,7 @@ static inline void B1_put_word(unsigned short base, unsigned int val)
 	B1_put_byte(base, (val >> 24) & 0xff);
 }
 
-static inline unsigned int B1_get_slice(unsigned short base,
+static inline unsigned int B1_get_slice(unsigned int base,
 					unsigned char *dp)
 {
 	unsigned int len, i;
@@ -342,7 +345,7 @@ static inline unsigned int B1_get_slice(unsigned short base,
 	return len;
 }
 
-static inline void B1_put_slice(unsigned short base,
+static inline void B1_put_slice(unsigned int base,
 				unsigned char *dp, unsigned int len)
 {
 	unsigned i = len;
@@ -375,7 +378,7 @@ static inline void B1_put_slice(unsigned short base,
 	}
 }
 
-static void b1_wr_reg(unsigned short base,
+static void b1_wr_reg(unsigned int base,
                       unsigned int reg,
 		      unsigned int value)
 {
@@ -384,7 +387,7 @@ static void b1_wr_reg(unsigned short base,
         B1_put_word(base, value);
 }
 
-static inline unsigned int b1_rd_reg(unsigned short base,
+static inline unsigned int b1_rd_reg(unsigned int base,
                                      unsigned int reg)
 {
 	B1_put_byte(base, READ_REGISTER);
@@ -393,14 +396,14 @@ static inline unsigned int b1_rd_reg(unsigned short base,
 	
 }
 
-static inline void b1_set_test_bit(unsigned short base,
+static inline void b1_set_test_bit(unsigned int base,
 				   int cardtype,
 				   int onoff)
 {
     b1_wr_reg(base, B1_STAT0(cardtype), onoff ? 0x21 : 0x20);
 }
 
-static inline int b1_get_test_bit(unsigned short base,
+static inline int b1_get_test_bit(unsigned int base,
                                   int cardtype)
 {
     return (b1_rd_reg(base, B1_STAT0(cardtype)) & 0x01) != 0;
@@ -482,7 +485,7 @@ int B1_valid_port(unsigned port, int cardtype)
    }
 }
 
-void B1_setinterrupt(unsigned short base,
+void B1_setinterrupt(unsigned int base,
 			         unsigned irq, int cardtype)
 {
 	switch (cardtype) {
@@ -500,17 +503,17 @@ void B1_setinterrupt(unsigned short base,
 	 }
 }
 
-unsigned char B1_disable_irq(unsigned short base)
+unsigned char B1_disable_irq(unsigned int base)
 {
 	return b1outp(base, B1_INSTAT, 0x00);
 }
 
-void T1_disable_irq(unsigned short base)
+void T1_disable_irq(unsigned int base)
 {
       t1outp(base, T1_IRQMASTER, 0x00);
 }
 
-void B1_reset(unsigned short base)
+void B1_reset(unsigned int base)
 {
 	b1outp(base, B1_RESET, 0);
 	udelay(55 * 2 * 1000);	/* 2 TIC's */
@@ -522,7 +525,7 @@ void B1_reset(unsigned short base)
 	udelay(55 * 2 * 1000);	/* 2 TIC's */
 }
 
-void T1_reset(unsigned short base)
+void T1_reset(unsigned int base)
 {
         /* reset T1 Controller */
         B1_reset(base);
@@ -534,7 +537,7 @@ void T1_reset(unsigned short base)
 	t1outp(base, T1_RESETBOARD, 0xf);
 }
 
-int B1_detect(unsigned short base, int cardtype)
+int B1_detect(unsigned int base, int cardtype)
 {
 	int onoff, i;
 
@@ -580,7 +583,7 @@ int B1_detect(unsigned short base, int cardtype)
 	return 0;
 }
 
-int T1_detectandinit(unsigned short base, unsigned irq, int cardnr)
+int T1_detectandinit(unsigned int base, unsigned irq, int cardnr)
 {
 	unsigned char cregs[8];
 	unsigned char reverse_cardnr;
@@ -652,7 +655,7 @@ int T1_detectandinit(unsigned short base, unsigned irq, int cardnr)
 
 extern int loaddebug;
 
-int B1_load_t4file(unsigned short base, avmb1_t4file * t4file)
+int B1_load_t4file(unsigned int base, avmb1_t4file * t4file)
 {
 	/*
 	 * Data is in user space !!!
@@ -691,7 +694,7 @@ int B1_load_t4file(unsigned short base, avmb1_t4file * t4file)
 	return 0;
 }
 
-int B1_load_config(unsigned short base, avmb1_t4file * config)
+int B1_load_config(unsigned int base, avmb1_t4file * config)
 {
 	/*
 	 * Data is in user space !!!
@@ -747,7 +750,7 @@ int B1_load_config(unsigned short base, avmb1_t4file * config)
 	return 0;
 }
 
-int B1_loaded(unsigned short base)
+int B1_loaded(unsigned int base)
 {
 	int i;
 	unsigned char ans;
@@ -797,7 +800,7 @@ static inline void parse_version(avmb1_card * card)
  * ------------------------------------------------------------------- 
  */
 
-void B1_send_init(unsigned short port,
+void B1_send_init(unsigned int port,
 	     unsigned int napps, unsigned int nncci, unsigned int cardnr)
 {
 	unsigned long flags;
@@ -811,7 +814,7 @@ void B1_send_init(unsigned short port,
 	restore_flags(flags);
 }
 
-void B1_send_register(unsigned short port,
+void B1_send_register(unsigned int port,
 		      __u16 appid, __u32 nmsg,
 		      __u32 nb3conn, __u32 nb3blocks, __u32 b3bsize)
 {
@@ -828,7 +831,7 @@ void B1_send_register(unsigned short port,
 	restore_flags(flags);
 }
 
-void B1_send_release(unsigned short port,
+void B1_send_release(unsigned int port,
 		     __u16 appid)
 {
 	unsigned long flags;
@@ -840,7 +843,7 @@ void B1_send_release(unsigned short port,
 	restore_flags(flags);
 }
 
-void B1_send_message(unsigned short port, struct sk_buff *skb)
+void B1_send_message(unsigned int port, struct sk_buff *skb)
 {
 	unsigned long flags;
 	__u16 len = CAPIMSG_LEN(skb->data);
